--- simple-chat/src/chatlist.html
@@ -1,65 +1,0 @@
-<!DOCTYPE html>
-<html lang="ru">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Messenger</title>
-    <link rel="stylesheet"
-        href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
-</head>
-
-<body style="margin: 0">
-    <div class="container">
-        <div class="workflow">
-            <header class="header">
-                <div class="burger">
-                    <div class="burger-line"></div>
-                    <div class="burger-line"></div>
-                    <div class="burger-line"></div>
-                    <div class="popup-menu">
-                        <div class="popup-option">Профиль</div>
-                        <div class="popup-option">Настройки</div>
-                        <div class="popup-option">Выйти</div>
-                    </div>
-                </div>
-                <h2 class="app">Messenger</h2>
-                <div class="search">
-                    <i class="material-symbols-outlined toucheble icons" title="Поиск">search</i>
-                </div>
-            </header>
-            <div class="chat-list">
-            </div>
-            <div>
-                <div class="create-chats" title="Создать чат">
-                    <i class="material-symbols-outlined icon-plus">add</i>
-                </div>
-            </div>
-        </div>
-        <div class="newchat-menu">
-            <div class="chatmenu-header">
-                <h3 class="chatmenu-title">Создать чат</h3>
-                <div class="nocreate-chat">
-                    <i class="material-symbols-outlined chat-menu-btn" title="Отмена">close</i>
-                </div>
-            </div>
-            <div>
-                <input type="text" class="chatmenu-lines" id="chat-name" placeholder="Название чата">
-            </div>
-            <div>
-                <input type="file" id="avatar-input" accept="image/*">
-                <label for="avatar-input" class="avatar-label" title="Аватар чата">
-                    <img src="https://upload.wikimedia.org/wikipedia/commons/7/7c/Profile_avatar_placeholder_large.png"
-                        id="avatar-placeholder" class="avatar" alt="Аватар чата">
-                </label>
-            </div>
-            <div class="create-chat">
-                <button class="chat-menu-btn material-symbols-outlined toucheble" title="Подтвердить">
-                    check
-                </button>
-            </div>
-        </div>
-    </div>
-</body>
-
-</html>
