+++ simple-chat/src/components/chat/mocks.js
