+++ simple-chat/src/index.css
@@ -1,191 +1,2 @@
-.chat-container {
-    font-size: 15px;
-    font-family: Helvetica, Arial, sans-serif;
-    font-weight: 400;
-    line-height: 20px;
-    height: 100vh;
-    display: flex;
-    flex-direction: column;
-    background-color: #f5f5f5;
-}
-
-.chat-header {
-    display: flex;
-    align-items: center;
-    padding: 10px;
-    background-color: #9e63e9;
-    color: white;
-    position: sticky;
-    top: 0;
-    z-index: 1;
-}
-
-.chat-header .avatar {
-    width: 40px;
-    height: 40px;
-    border-radius: 50%;
-    margin-right: 10px;
-    margin-left: 10px;
-}
-
-.user-block .username {
-    margin: 0;
-}
-
-.chat-header .header-right {
-    display: flex;
-    align-items: center;
-    margin-left: auto;
-}
-
-.chat-header .header-right .icons {
-    font-size: 25px;
-    margin: 0 10px;
-}
-
-.user-block {
-    display: flex;
-    flex-direction: column;
-}
-
-.chat-body {
-    height: calc(100vh - 100px);
-    overflow-y: auto;
-    padding: 10px;
-}
-
-.chat-message {
-    display: flex;
-    margin-bottom: 10px;
-}
-
-.chat-message.user {
-    display: flex;
-    flex-direction: row-reverse;
-}
-
-.chat-message .message-content {
-    padding: 10px;
-    border-radius: 10px;
-    max-width: 80%;
-    word-wrap: break-word;
-    margin-left: 10px;
-}
-
-.chat-message.user .message-img {
-    padding: 4px;
-    border-radius: 10px;
-    max-width: 95%;
-    word-wrap: break-word;
-    margin-left: 10px;
-    background-color: #c398ff;
-    color: white;
-}
-
-.chat-message.other .message-img {
-    padding: 4px;
-    border-radius: 10px;
-    max-width: 95%;
-    word-wrap: break-word;
-    margin-left: 10px;
-    background-color: #e0e0e0;
-}
-
-.chat-message.user .message-content {
-    background-color: #bb8cfdd3;
-    color: white;
-}
-
-.chat-message.other .message-content {
-    background-color: #e0e0e0;
-}
-
-.chat-message .message-time {
-    font-size: 12px;
-    color: #5d5d5d;
-    display: flex;
-    flex-direction: row-reverse;
-}
-
-.img-time {
-    padding-right: 5px;
-}
-
-.img {
-    border-radius: 3px;
-}
-
-.chat-footer {
-    display: flex;
-    align-items: center;
-    padding: 10px;
-    background-color: white;
-    position: sticky;
-    bottom: 0;
-    z-index: 1;
-}
-
-.chat-footer .form-input {
-    flex: 1;
-    padding: 10px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-}
-
-.chat-footer .form-input:focus {
-    outline-color: #818181;
-}
-
-.material-symbols-outlined.checked {
-    color: #2f0369;
-    font-size: 18px;
-    font-variation-settings:
-        'FILL' 0,
-        'wght' 300,
-        'GRAD' 200,
-        'opsz' 48
-}
-
-.material-symbols-outlined.send {
-    padding-top: 3px;
-    margin-left: 5px;
-    margin-right: 3px;
-    color: #9e63e9;
-    font-size: 40px;
-    font-variation-settings:
-        'FILL' 1,
-        'wght' 100,
-        'GRAD' 200,
-        'opsz' 48
-}
-
-.material-symbols-outlined.back {
-    font-size: 25px;
-    padding: 5px;
-    margin-right: 5px;
-    margin-left: 10px;
-}
-
-.material-symbols-outlined.attach {
-    color: #ccc;
-    font-size: 25px;
-    font-variation-settings:
-        'FILL' 0,
-        'wght' 200,
-        'GRAD' 0,
-        'opsz' 40
-}
-
-.material-symbols-outlined.attach:hover {
-    color: #818181;
-    cursor: pointer;
-    transition: transform 0.2s ease;
-    transform: translateY(-3px);
-}
-
-.material-symbols-outlined.toucheble:hover {
-    cursor: pointer;
-    color: #2f0369;
-    transition: transform 0.2s ease;
-    transform: translateY(-3px);
-}
+@import "components/chat/chat.css";
+@import "components/chatlist/chatlist.css";
