+++ simple-chat/src/index.html
@@ -4,37 +4,12 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Chat - Actions</title>
+    <title>Simple-Chat</title>
     <link rel="stylesheet"
         href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
 </head>
 
-<body style="margin: 0">
-    <div class="chat-container">
-        <div class="chat-header">
-            <i class="material-symbols-outlined back toucheble" title="Назад" onclick="window.location.href='chatlist.html'">arrow_back_ios_new</i>
-            <img class="avatar" src="https://avatar.iran.liara.run/public/98" alt="Аватар">
-            <div class="user-block">
-                <h2 class="username">Дженнифер</h2>
-                <span>в сети</span>
-            </div>
-            <div class="header-right">
-                <i class="material-symbols-outlined toucheble icons" title="Поиск">search</i>
-                <i class="material-symbols-outlined toucheble icons" title="Дополнительно">more_vert</i>
-            </div>
-        </div>
-        <div class="chat-body">
-        </div>
-        <form class="form">
-            <div class="chat-footer">
-                <input class="form-input" name="message-text" type="text" placeholder="Сообщение">
-                <i class="material-symbols-outlined attach" title="Прикрепить файл">attach_file</i>
-                <div class="submit">
-                    <i class="material-symbols-outlined send toucheble" title="Отправить">send</i>
-                </div>
-            </div>
-        </form>
-    </div>
+<body style="margin: 0" id="root">
 </body>
 
 </html>
